+++ hart/master.py
@@ -5,6 +5,7 @@ import yaml
 
 from . import utils
 from .constants import DEBIAN_VERSIONS
+from .exceptions import UserError
 from .ssh import get_verified_ssh_client, ssh_run_command, ssh_run_init_script
 
 
@@ -21,6 +22,7 @@ def create_master(
         grains=None,
         script=None,
         authorize_key=None,
+        connect_via_private_ip=False,
         **kwargs
         ):
     hart_node = create_master_node(
@@ -34,6 +36,7 @@ def create_master(
         private_networking,
         minion_config,
         grains,
+        connect_via_private_ip=connect_via_private_ip,
         **kwargs
     )
     try:
@@ -55,6 +58,7 @@ def create_master_node(
         private_networking=False,
         minion_config=None,
         grains=None,
+        connect_via_private_ip=False,
         **kwargs
         ):
     ssh_canary = utils.create_token()
@@ -97,8 +101,17 @@ def create_master_node(
                 **kwargs)
             node = provider.wait_for_public_ip(node)
             public_ip = node.public_ips[0]
-            print('Master running at %s' % public_ip)
-            return utils.HartNode(minion_id, public_ip, node, provider, ssh_key, ssh_canary, extra)
+            connect_ip = public_ip
+            if connect_via_private_ip:
+                if not node.private_ips:
+                    raise UserError('Cannot connect via private IP, node has none')
+                connect_ip = node.private_ips[0]
+                print('Master running at %s, connecting via private IP %s' % (
+                    public_ip, connect_ip))
+            else:
+                print('Master running at %s' % public_ip)
+            return utils.HartNode(minion_id, public_ip, node, provider, ssh_key,
+                ssh_canary, extra, connect_ip)
         except:
             traceback.print_exc()
             if node:
@@ -109,8 +122,9 @@ def create_master_node(
 
 def connect_to_master(hart_node, script, authorize_key=None):
     username = hart_node.provider.username
+    connect_ip = hart_node.connect_ip or hart_node.public_ip
     with get_verified_ssh_client(
-            hart_node.public_ip,
+            connect_ip,
             hart_node.ssh_key,
             hart_node.ssh_canary,
             username) as client: